Change function signature for binary download to make it explicit whether file is compressed (#31)

+++ src/commands/utils/helper/binaryDownloadHelper.ts
@@ -17,13 +17,31 @@ function getToolBinaryFolder(toolName: string, version: string): string {
 function getToolDownloadFolder(toolName: string): string {
     return path.join(getToolBaseInstallFolder(toolName), "download");
 }
- 
+
+type CommonDownloadSpec = {
+    downloadUrl: string
+}
+
+export type BinaryDownloadSpec = CommonDownloadSpec & {
+    isCompressed: false
+};
+
+export type ArchiveDownloadSpec = CommonDownloadSpec & {
+    isCompressed: true,
+    pathToBinaryInArchive: string
+};
+
+export type DownloadSpec = BinaryDownloadSpec | ArchiveDownloadSpec;
+
+function isArchive(downloadSpec: DownloadSpec): downloadSpec is ArchiveDownloadSpec {
+    return downloadSpec.isCompressed;
+}
+
 export async function getToolBinaryPath(
     toolName: string,
     version: string,
-    downloadUrl: string,
-    pathToBinaryInArchive: string | undefined,
     binaryFilename: string,
+    downloadSpec: DownloadSpec
     ): Promise<Errorable<string>> {
 
     const binaryFolder = getToolBinaryFolder(toolName, version);
@@ -32,71 +50,53 @@ export async function getToolBinaryPath(
     if (fs.existsSync(binaryFilePath)) {
        return {succeeded: true, result: binaryFilePath};
     }
-    if (pathToBinaryInArchive === undefined) {
-        return await longRunning(`Downloading ${toolName} to ${binaryFilePath}.`, () => downloadBinaryWithoutArchive(toolName, binaryFilePath, downloadUrl));
-    }
-    return await longRunning(`Downloading ${toolName} to ${binaryFilePath}.`, () => downloadBinary(toolName, binaryFilePath, downloadUrl, pathToBinaryInArchive));
+
+    return await longRunning(`Downloading ${toolName} to ${binaryFilePath}.`, () => downloadTool(toolName, binaryFilePath, downloadSpec));
 }
 
-async function downloadBinary(
+async function downloadTool(
     toolName: string,
     binaryFilePath: string,
-    downloadUrl: string,
-    pathToBinaryInArchive: string
+    downloadSpec: DownloadSpec
 ): Promise<Errorable<string>> {
 
-    const downloadFileName = downloadUrl.substring(downloadUrl.lastIndexOf('/') + 1);
+    const downloadFileName = downloadSpec.downloadUrl.substring(downloadSpec.downloadUrl.lastIndexOf('/') + 1);
     const downloadFolder = getToolDownloadFolder(toolName);
     const downloadFilePath = path.join(downloadFolder, downloadFileName);
 
-    const downloadResult = await download.once(downloadUrl, downloadFilePath);
+    const downloadResult = await download.once(downloadSpec.downloadUrl, downloadFilePath);
     if (failed(downloadResult)) {
         return {
             succeeded: false,
-            error: `Failed to download binary from ${downloadUrl}: ${downloadResult.error}`
-        };
-    }
-    const decompress = require("decompress");
-
-    try {
-        await decompress(downloadFilePath, downloadFolder);
-    } catch (error) {
-        return {
-            succeeded: false,
-            error: `Failed to unzip binary ${downloadFilePath} to ${downloadFolder}: ${error}`
+            error: `Failed to download binary from ${downloadSpec.downloadUrl}: ${downloadResult.error}`
         };
     }
 
-    // Remove zip.
-    fs.unlinkSync(downloadFilePath);
+    if (isArchive(downloadSpec)) {
+        const decompress = require("decompress");
 
-    // Avoid `download.once()` thinking that the zip file is already downloaded the next time.
+        try {
+            await decompress(downloadFilePath, downloadFolder);
+        } catch (error) {
+            return {
+                succeeded: false,
+                error: `Failed to unzip binary ${downloadFilePath} to ${downloadFolder}: ${error}`
+            };
+        }
+    
+        // Remove zip.
+        fs.unlinkSync(downloadFilePath);
+
+        // Move extracted binary to where we want it.
+        const unzippedBinaryFilePath = path.join(downloadFolder, downloadSpec.pathToBinaryInArchive);
+        await moveFile(unzippedBinaryFilePath, binaryFilePath);
+    } else {
+        await moveFile(downloadFilePath, binaryFilePath);
+    }
+
+    // Avoid `download.once()` thinking that the downloaded file is already downloaded the next time.
     // If there's any failure after this, we *want* it to be downloaded again.
     download.clear(downloadFilePath);
-
-    // Move file to more flatten structure.
-    const unzippedBinaryFilePath = path.join(downloadFolder, pathToBinaryInArchive);
-
-    await moveFile(unzippedBinaryFilePath, binaryFilePath);
-
-    //If linux check -- make chmod 0755
-    fs.chmodSync(path.join(binaryFilePath), '0755');
-    return { succeeded: true, result: binaryFilePath };
-}
-
-async function downloadBinaryWithoutArchive(
-    toolName: string,
-    binaryFilePath: string,
-    downloadUrl: string
-): Promise<Errorable<string>> {
-
-    const downloadResult = await download.once(downloadUrl, binaryFilePath);
-    if (failed(downloadResult)) {
-        return {
-            succeeded: false,
-            error: `Failed to download binary from ${downloadUrl}: ${downloadResult.error}`
-        };
-    }
 
     //If linux check -- make chmod 0755
     fs.chmodSync(path.join(binaryFilePath), '0755');
